dom/tests/mochitest/mouselock/test_lockLostCallBack.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #test-element { background-color: #94E01B; width:100px; height:100px}
    #test-element-two { background-color: #000; width:100px; height:100px}
  </style>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
  <div id="content">
    <div id="test-element-two"></div>
    <div id="test-element" onclick="test();"></div>
  </div>
  <pre id="test">
    <script type="application/javascript">
      /** Test for Bug 633602 **/
      SimpleTest.waitForExplicitFinish()
      SimpleTest.waitForFocus(function() {
        SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

        var pointer = navigator.pointer;

        var testElement = document.getElementById("test-element");
        var testElementTwo = document.getElementById("test-element-two");

        var i = 0;
        var eventCount = 0;

        var testFunction = function(){
          if (document.mozFullScreenElement === testElementTwo) {
            pointer.lock(testElementTwo, function(){
              is (i, 1, "CallBack function on first lock lost did not fire");
              document.mozCancelFullScreen();
              SimpleTest.finish();
              }, function(){ });
          }else
           setTimeout(testFunction, 10);
        }

        var listener = (function (e) {  
          eventCount++;
          if (eventCount == 1)
          {  
            pointer.lock(testElement, function(e){ testElementTwo.mozRequestFullScreen(); setTimeout(testFunction, 10);}, function(e){ i++;});
          }
        });

        document.addEventListener("mozfullscreenchange", listener,false);
        testElement.mozRequestFullScreen();
      });
    </script>
  </pre>
</body>
</html>
